Refactor: Move Randomizer to utils package
Move Randomizer interface and implementations to a new utils package.
