+++ backend/src/main/java/com/odde/doughnut/entities/Notebook.java
@@ -6,8 +6,8 @@ import static com.theokanning.openai.service.OpenAiService.defaultObjectMapper;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.odde.doughnut.exceptions.ApiException;
-import com.odde.doughnut.models.Randomizer;
 import com.odde.doughnut.services.graphRAG.BareNote;
+import com.odde.doughnut.utils.Randomizer;
 import jakarta.persistence.*;
 import java.sql.Timestamp;
 import java.util.ArrayList;
