+++ backend/src/main/java/com/odde/doughnut/utils/Randomizer.java
@@ -1,4 +1,4 @@
-package com.odde.doughnut.models;
+package com.odde.doughnut.utils;
 
 import java.util.List;
 import java.util.Optional;
