+++ backend/src/main/java/com/odde/doughnut/utils/randomizers/NonRandomizer.java
@@ -1,7 +1,7 @@
-package com.odde.doughnut.models.randomizers;
+package com.odde.doughnut.utils.randomizers;
 
 import com.odde.doughnut.controllers.dto.Randomization;
-import com.odde.doughnut.models.Randomizer;
+import com.odde.doughnut.utils.Randomizer;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
