+++ backend/src/test/java/com/odde/doughnut/services/ai/AiQuestionGeneratorTests.java
@@ -7,11 +7,11 @@ import static org.mockito.Mockito.*;
 import com.odde.doughnut.configs.ObjectMapperConfig;
 import com.odde.doughnut.entities.Note;
 import com.odde.doughnut.factoryServices.ModelFactoryService;
-import com.odde.doughnut.models.Randomizer;
-import com.odde.doughnut.models.randomizers.RealRandomizer;
 import com.odde.doughnut.services.GlobalSettingsService;
 import com.odde.doughnut.testability.MakeMe;
 import com.odde.doughnut.testability.OpenAIChatCompletionMock;
+import com.odde.doughnut.utils.Randomizer;
+import com.odde.doughnut.utils.randomizers.RealRandomizer;
 import com.theokanning.openai.client.OpenAiApi;
 import java.util.Arrays;
 import java.util.List;
